Add room power levels to verify user in room response
If Synapse has the related admin api (https://github.com/matrix-org/synapse/pull/9168), fetch power levels and add them to the result of the verify user in room.

The response structure is the m.room.power_levels state event content with the users stripped out. The user level is found separately on the higher level to avoid confusion related to all the power levels for users in the room.