Refactor OpenID check behaviour
Default behaviour has been changed to verify any OpenID tokens of any homeserver. This keeps the code, configuration and examples simpler by reversing to requiring configuring a single homeserver for OpenID verifications explicitly.

Additionally this allows for the server and client Matrix API's to be on different domains, which was not true before when running against a single homeserver.

This is a breaking change which has been highlighted in the changelog.

This PR also fixes verifying tokens against homeservers which use SRV to delegate to another domain.